Extracted inner class into a separate class
